requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==5.0.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-beat==2.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
lxml==4.6.2
plyer==2.0.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
